auto pow fix 2

+++ MonitorClient/status/views.py
@@ -70,21 +70,22 @@ def autopow(request):
             print(onjob)
             print(offjob)
         except:
-            pass        
-    try:
-        cron.remove_all(comment='on')
-        cron.remove_all(comment='off')
-    except:
-        pass
-    on_hour = request.POST.get('on').split(":")[0]
-    on_minute = request.POST.get('on').split(":")[1]
-    off_hour = request.POST.get('off').split(":")[0]
-    off_minute = request.POST.get('off').split(":")[1]
-    onjob = cron.new(command = 'echo "on 0" | cec-client -s -d 1; wget http://%s:8000/status/switch/1' % IP, comment = "on")
-    onjob.hour.on(on_hour)
-    onjob.minute.also.on(on_minute)
-    offjob = cron.new(command = 'echo "standby 0" | cec-client -s -d 1; wget http://%s:8000/status/switch/1' % IP, comment = "off")
-    offjob.hour.on(off_hour)
-    offjob.minute.also.on(off_minute)
-    cron.write()
+            pass
+    elif request.POST:      
+        try:
+            cron.remove_all(comment='on')
+            cron.remove_all(comment='off')
+        except:
+            pass
+        on_hour = request.POST.get('on').split(":")[0]
+        on_minute = request.POST.get('on').split(":")[1]
+        off_hour = request.POST.get('off').split(":")[0]
+        off_minute = request.POST.get('off').split(":")[1]
+        onjob = cron.new(command = 'echo "on 0" | cec-client -s -d 1; wget http://%s:8000/status/switch/1' % IP, comment = "on")
+        onjob.hour.on(on_hour)
+        onjob.minute.also.on(on_minute)
+        offjob = cron.new(command = 'echo "standby 0" | cec-client -s -d 1; wget http://%s:8000/status/switch/1' % IP, comment = "off")
+        offjob.hour.on(off_hour)
+        offjob.minute.also.on(off_minute)
+        cron.write()
     return HttpResponse("done")
